cp -f: don't unlink to destination file unless we fail to open it
According to section 3.a.iii of POSIX's description of cp, the existing
destination file should only be unlinked if we fail to open it with
O_WRONLY | O_TRUNC. Currently, it is unconditionally unlinked, which
leads to surprising behaviour when the destination file is a symbolic
link. Other cp(1) implementations, including NetBSD, OS X and GNU,
follow POSIX in handling this case.