cp: handle copies to dangling symlinks when -RP is specified
Before performing a copy of a source file, we stat(2) the target to
determine whether or not it already exists. However, when -RP is
specified, symlinks are not to be followed, but copied instead. The
check for an existing target is therefore incorrect when a dangling
symlink is present in the source tree, and can result in spurious
failures if a dangling symlink is already present in the target tree.

Additionally, the previous behaviour would generate spurious failures if
the target was a symlink to a directory, since S_ISDIR checks would be
incorrectly applied to the target of the symlink rather than the symlink
itself.