smp_rendezvous: provide a trap frame to rendezvous handlers
On architectures where IPI handlers are invoked as interrupt filters,
the trap frame is already available since the generic interrupt handler
sets td_intr_frame before calling the filter.

On i386 and amd64, smp_rendezvous_action() is called directly from the
interrupt handler; a pointer to the trap frame can be passed as an
argument. On sparc64, the trap frame is already passed as an argument,
so no changes are necessary.