Add stack_save_td_running(9).
This function allows the caller to capture the kernel stack of a running
thread. It is currently only implemented on i386 and amd64; on these
architectures, it is implemented by raising an NMI on the CPU on which
the target thread is currently running. Unlike stack_save_td(9), it may
fail, e.g. if the target thread is running in user mode.

Have kern.proc.kstack use this function for running threads.