Make sched_foreach_on_runq() pass a thread pointer to its callback and provide an additional argument from the caller.
Note for later: the API is not great since it assumes that scheduler
run queues and CPUs are 1-1. It's probably better (and simpler) to have
a scheduler function that iterates over all queues for the caller; there
doesn't seem to be any necessity to be running on the CPU local to the
run queue.