strengthen lock assertions in callout_drain()
r277723 causes a warning to be printed if a non-sleepable lock is held
when calling callout_drain(). However, it excludes the callout's lock
from this check, which is misleading since
1. it's also incorrect to hold the callout lock when calling
   callout_drain(), and
2. callout locks are always non-sleepable anyway.

This change includes c->c_lock in the witness check, and additionally
asserts that the callout lock is not held by a caller of
callout_drain().