Restore the previous interrupt frame after exiting the critical section.
This is a hack to ensure that the interrupt frame is available if the
thread is preempted by critical_exit(). I don't *think* it'll cause any
problems, but it represents a flaw in the approach. It seems that there
are several races inherent in using the td_intr_frame field of a
preempted thread; a more robust solution may involve modifying the iret
path to check and fixup the PC before a return to kernel mode.