Rework the proc_readmem/writemem interface per kib's suggestion.
Have the new subroutines return the number of bytes read or written
rather than a binary success/failure status. If an error occurs, i.e.
proc_rwmem() returned a non-zero status and no bytes were read or
written, we return -1.

The behaviour of existing callers is preserved. That is, in most cases
we return an error if a short read or write occurs.

proc_read_string() is simplified somewhat: as the comment notes, we may
get a short read if one of the requested chunks falls in an unmapped
page. We now return success in this case rather than resorting to
single-byte reads.

This change also simplifies handling of PT_READ_* and PT_WRITE_*, per
jhb.