Simplify spin counting for lockstat probes.
We only use the spin count to determine whether the <lock>-spin probe
should fire. It's supposed to fire if we've spent any time adaptively
spinning for the lock, so only increment it immediately before we're
about to do so (and not in the spin loop itself). We also don't need to
keep track of the "sleep count".

While here, fix some style bugs.