Add PQ_STASIS.
The is a paging queue used to store dirty inactive anon pages that cannot
be laundered. Currently this occurs when no swap devices are
configured, but could be used to handle other pager errors.

This change adds swapon and swapoff EVENTHANDLERs, serialized by the
swapconf lock. The laundry thread's handlers for these events increment
and decrement a count of configured swap devices. If the number of swap
devices is 0, an error of VM_PAGER_FAIL for a putpages call on a dirty
page belonging to an object of type DEFAULT or SWAP results in the page
entering the stasis queue. If the number of configured swap devices
transitions from 0 to 1, all pages in the stasis queue are placed at the
end of the laundry queue.

There is no limit on the number of pages in stasis. pagedaemon targets
are not scaled according to the number of pages in stasis. This queue's
purpose is to ensure that the pagedaemon will not waste CPU time
examining pages it cannot reclaim.

This change modifies top(1) and vmstat(8) to display the number of
changes in stasis.