Activate triggered kevents when they are enabled
r274560 modified kevent_register() to avoid testing knotes for disabled
kevents, presumably as a micro-optimization. However, this check is done
before possibly enabling the disabled kevent with EV_ENABLE, which means
that the knote will not be activated by the kevent_register() call even
though the event is enabled when this call returns.

Note that a knote for disabled kevent may be activated independently of
a kevent_register() call; the problem occurs only when the event trigger
condition does not undergo a level transition while the event is
disabled.

This change also modifies an existing kqueue test to exercise this case.