Implement a working set algorithm for zone bucket caches
UMA implements two levels of caching. At the first level, we keep items
in per-CPU buckets. The second level is the zone bucket list, which is
protected by the zone mutex. Under memory pressure, uma_reclaim() is
periodically invoked in order to reclaim memory from the bucket cache
and full slabs. When free memory is very scarce (free_count < free_min)
or when the kmem KVA arena is exhausted, we also reclaim the per-CPU
caches. This should never happen during normal operation of the system.

Without this change, uma_reclaim() frees the entire bucket cache. This
ensures that UMA zones don't cache excessive amounts of memory, but is
too aggressive on zones with a working set that exceeds the capacity of
its per-CPU caches. This change modifies uma_reclaim to free memory more
intelligently.

This commit consists of several parts:
- Bucket cache accounting. We now keep track of the total number of
  zone items in the bucket cache.
- UMA callout updates. The callout now runs every lowmem_period+1
  seconds and uses the bucket cache size to estimate the current working
  set size for the bucket cache (as opposed to the entire zone). The
  estimate is a decaying average over 2 minutes; the estimate is allowed
  to increase more quickly than it decays. This means that the estimate
  grows quickly in the face of activity bursts.
- uma_reclaim() changes. Under memory pressure, we now prune the zone's
  bucket cache according to the estimated working set size. When under
  severe memory pressure, we discard the per-CPU caches as before.
  Buckets are reclaimed in LRU order.

Discussed with:	jeff