Fix memguard(9) when INVARIANTS is enabled.
With r284861, zones use the trash ctor and dtor by default. This is
incompatible with memguard. Modify the UMA debug functions to be
no-ops if the item was allocated from memguard. This also fixes ctors
such as mb_ctor_pack(), which invokes the trash ctor in addition to
performing some initialization.