Don't preserve the page object/pindex in vm_page_insert_after()
Per the KASSERT at the beginning of the function, we expect that the
page does not belong to any object, so its object and pindex fields are
meaningless. Reset them in the rare case that vm_radix_insert() fails.

I note that we do not clear the pindex field when freeing a page:
vm_page_remove() only clears the object field.