Don't deliver gmirror BIOs during provider teardown
This results in lock recursion if any regular or sync BIOs are on
sc_queue when the provider is destroyed, for example when the last
component of a mirror fails. g_mirror_done() and g_mirror_sync_done()
both acquire the queue lock, but we're already holding it in order to
drain the queue. Instead, just free the BIOs.