Merge {amd64,i386}/instr_size.c into x86/instr_size.c.
Also reduce the diff between us and upstream: the input model will
always be DATAMODEL_NATIVE because of a bug (p_model is never set but
always initialized to 0). DATAMODEL_NATIVE == DATAMODEL_LP64 on amd64
and DATAMODEL_NATIVE != DATAMODEL_LP64 on i386, so the upstream code
already does the right thing.