Avoid sleeping after a failed vm_page_alloc_contig() call during boot.
After r297466, a VM_WAIT during boot will result in a hang since the
pagedaemon may not yet be running. Previously, VM_WAIT would just
return immediately.

I tried to address one instance of this issue with r302147, but there
are others in vesa(4), for instance. It seems to me that we should just
avoid the VM_WAIT if the kernel is still booting to avoid these types
of regressions after r297466.