Fix swap pager readahead after r292373.
The removal of vm_fault_additional_pages() means that a hard fault on a
OBJT_SWAP-backed page will cause exactly one page to be read in. This
change restores readahead and readbehind for the swap pager in
swap_pager_getpages(). haspages() is modified to return the largest
contiguous non-resident range of pages in the swap block containing the
requested range.