Don't fail disk state change if g_mirror_clear_metadata() fails.
We may fail to clear the metadata if we attempt to remove the disk from
the mirror as it is being destroyed, e.g. because of a concurrent hot
swap. This is not a fatal error since a failure from
g_mirror_clear_metadata() will result in the mirror's generation ID
being bumped, so the disk in question will not be added back into the
mirror upon a retaste.