Remove a lock assertion in vm_reserv_free_page().
This assertion was incorrect in the case where some but not all of a
reservations pages were renamed. In particular, it's not the case that
m->object == rv->object for every page in an active reservation.

However, we can still locklessly check whether rv->object == NULL:
rv->object can't transition to NULL until all pages from the reservation
have been freed.