Lock the prefix list and add refcounting for prefixes.
This change extends the nd6 lock to protect the NDP prefix list and the
list of advertising routers associated with each prefix. To handle cases
where the nd6 lock must be dropped while iterating over either the
prefix or default router lists, a generation counter is used to track
modifications to the lists. Additionally, a mutex is used to serialize
prefix transitions between on-link and off-link.