Batch insertions of pages into VM page queues.
Currently, the pagequeue and page locks are both required to insert a
page into a page queue. This inhibits concurrency by serializing common
operations (page faults, unwirings) through the per-domain pagequeue
locks.

The page lock is acquired before the pagequeue lock. This change makes
use of that fact and adds per-page lock "batch queues." The enqueue
operation is modified so that the page is first inserted into its
corresponding batch queue. If the batch queue length is greater than a
small bound, the queue's contents are concatenated onto the main page
queue with an O(1) operation. Only the concatenation requires the queue
lock, so this approach increases concurrency at the expense of strict
LRU.

One complication is the possibility that a thread may attempt to dequeue
a page while it is still in its batch queue. This can be addressed by
adding a new page flag that is set upon insertion into a batch queue and
cleared when pages are pushed to the main queue. However, this turns the
concatenation into an O(n) operation with the page lock held. The
approach taken by this change is to instead always concatenate the batch
queue for a page prior to removal. This ensures that the page is not in
a batch queue and allows for O(1) concatenations at the cost of more
expensive enqueue and dequeue operations.

The batch queue bound is a function of the amount of system memory. It
is chosen such that no more than 0.1% of all the pages protected by a
given page lock can be enqueued in a batch queue simultaneously. This
gives a rough bound of 1 page per GB of memory.

TODO:
- vm_page_dequeue should only modify the queue length once
- can the pagedaemon avoid concatenating batch queues when freeing
  pages?