Shift some of the EKCD funcs around to keep them grouped together.
Now that all of the kerneldumpcrypto functions are called only in
kern_shutdown.c, this is more reasonable. This also lets us reduce the
number of ifdefs.

Also Remove the keysize parameter from mkdumpheader() and instead use
the configured dumper to get the key size. For textdumps we already
"hide" the configured crypto parameters by temporarily unsetting
dumper.kdc.