Convert a witness assertion into a warning
When a lock departs and is enrolled again, we panic if the new instance
does not belong to the same lock class as the old instance. However,
there are legitimate cases when this may occur while doing kernel
development. In particular, one may unload a kernel module, modify the
lock class used for one of the locks in the module, and load it again.
Witness can otherwise handle this, so downgrade the assertion to a
warning.